Use Timer in Timestamp()
--HG--
branch : issue#36

+++ src/client/sys.hpp
@@ -5,11 +5,11 @@
 
 namespace sys {
 
-/// Returns current timestamp in seconds.
-double Timestamp();
+// Returns current timestamp in ms.
+int64_t Timestamp();
 
 void Warning(const char* fmt, ...);
 
 } // namespace sys
 
-#endif  // BLOWMORPH_CLIENT_SYS_HPP_
+#endif // BLOWMORPH_CLIENT_SYS_HPP_
